refactor: update event bus keys and adjust frequency parameters in state machines

+++ control-unit-backend/project/src/main/java/com/cub/MqttAgentVerticle.java
@@ -53,7 +53,7 @@ public class MqttAgentVerticle extends AbstractVerticle {
             if (client != null && client.isConnected()) {
                 client.publish(
                         FREQ_TOPIC,
-                        Buffer.buffer(String.valueOf((Long) message.body())),
+                        Buffer.buffer(String.valueOf((int) message.body())),
                         MqttQoS.AT_LEAST_ONCE,
                         false,
                         false,
